[#1793] Split existing FFA constraints into witness and constraints files
