+++ msm/src/ffa/mod.rs
@@ -1,4 +1,4 @@
 pub mod columns;
 pub mod constraint;
 //pub mod interpreter;
-//pub mod witness;
+pub mod witness;
