+++ msm/src/main.rs
@@ -4,7 +4,9 @@ use kimchi::circuits::domains::EvaluationDomains;
 use poly_commitment::pairing_proof::PairingSRS;
 
 use kimchi_msm::columns::Column;
-use kimchi_msm::ffa::{columns::FFA_N_COLUMNS, constraint::MSMCircuitEnv};
+use kimchi_msm::ffa::{
+    columns::FFA_N_COLUMNS, constraint::get_exprs_add, witness::WitnessBuilder as FFAWitnessBuilder,
+};
 use kimchi_msm::lookups::LookupTableIDs;
 use kimchi_msm::precomputed_srs::get_bn254_srs;
 use kimchi_msm::prover::prove;
@@ -13,8 +15,8 @@ use kimchi_msm::{
     BN254G1Affine, BaseSponge, Ff1, Fp, OpeningProof, ScalarSponge, BN254, DOMAIN_SIZE,
 };
 
-pub fn generate_random_msm_witness() -> MSMCircuitEnv<BN254G1Affine> {
-    let mut circuit_env = MSMCircuitEnv::<BN254G1Affine>::empty();
+pub fn generate_random_msm_witness() -> FFAWitnessBuilder<BN254G1Affine> {
+    let mut circuit_env = FFAWitnessBuilder::<BN254G1Affine>::empty();
     let mut rng = thread_rng();
 
     let row_num = DOMAIN_SIZE;
@@ -40,7 +42,7 @@ pub fn main() {
 
     let circuit_env = generate_random_msm_witness();
     let proof_inputs = circuit_env.get_witness();
-    let constraint_exprs = circuit_env.get_exprs_add();
+    let constraint_exprs = get_exprs_add();
 
     println!("Proof inputs: {:?}", proof_inputs);
     println!("Constraints: {:?}", constraint_exprs);
